// src/models/files-properties-sofiral.ts
export interface FilesPropertiesSofiral {
  CodeClient: string;
  DateDebutFacture: string;
  DateFacture: string;
  DateFinFacture: string;
  Duplicata: boolean;
  Montant: number;
  NomClient: string;
  NomDossier: string;
  NomSociete: string;
  Nommage: string;
  NumFacture: string;
  Periodicite: string;
  id: string;
  isFile: boolean;
  name: string;
}